Refactor the classpaths
