--- botsing-model-generation/src/main/java/eu/stamp/botsing_model_generation/generation/behavioral_model/model/Model.java
@@ -1,4 +1,0 @@
-package eu.stamp.botsing_model_generation.generation.behavioral_model.model;
-
-public class Model {
-}
